requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.4.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cssselect==1.1.0
cycler==0.10.0
Django==3.2.7
django-celery-results==2.2.0
django-redis==5.0.0
djangorestframework==3.12.4
fake-useragent==0.1.11
idna==3.2
importlib-metadata==4.8.1
joblib==1.0.1
kiwisolver==1.3.2
kombu==5.1.0
lxml==4.6.3
matplotlib==3.4.3
mysqlclient==2.0.3
numpy==1.21.2
opencv-contrib-python==4.5.3.56
pandas==1.3.3
parse==1.19.0
Pillow==8.3.2
prompt-toolkit==3.0.20
pyee==8.2.2
pyparsing==2.4.7
pyppeteer==0.2.6
pyquery==1.4.3
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-html==0.10.0
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
sqlparse==0.4.2
threadpoolctl==2.2.0
tqdm==4.62.3
urllib3==1.26.7
uWSGI @ file:///home/conda/feedstock_root/build_artifacts/uwsgi_1614688744971/work
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
websockets==9.1
zipp==3.5.0
